Refactor rows of feature-table into a new feature-row class.
